Move successMessage to AddBook, adjust test
